refactor: main 메서드 속 실행 메시지 BaseballGame의 run 메서드로 분리
