--- src/main/java/baseball/NumberScanner.java
@@ -1,30 +1,0 @@
-package baseball;
-
-import java.util.ArrayList;
-import java.util.List;
-import camp.nextstep.edu.missionutils.Console;
-
-public class NumberScanner {
-
-    public static List<Integer> getNumbers() {
-        List<Integer> userNumbers = new ArrayList<>();
-        System.out.printf("숫자를 입력해주세요: ");
-        String input = Console.readLine();
-
-        if (input.length() != 3) {
-            throw new IllegalArgumentException();
-        }
-
-        for (char c : input.toCharArray()) {
-            if (c < 49 || c > 57) {
-                throw new IllegalArgumentException();
-            }
-            if (userNumbers.contains(Character.getNumericValue(c))) {
-                throw new IllegalArgumentException();
-            }
-            userNumbers.add(Character.getNumericValue(c));
-        }
-
-        return userNumbers;
-    }
-}
